[refactor] Extract XSD 1.1 store-time validation infrastructure out of MutableCollection
MutableCollection carried a schema cache (plain ConcurrentHashMap) plus the
namespace-resolution/validation logic that uses it -- neither is Collection
state. Splits this into two package-private classes:

- Xsd11SchemaCache: the per-namespace "needs XSD 1.1?" cache, now backed by
  Caffeine for consistency with the codebase's other caches (no eviction
  policy, since the namespace set is finite and admin-controlled -- this was
  never a correctness concern, just a style/testability one).
- Xsd11ValidationHelper: resolveXsd11SchemaForNamespace, the dynamic-discovery
  schema holder, parseOrValidateXmlSource, validateWithXsd11Schema, and the
  Xsd11LexicalHandlerForwarder record -- store-time XSD 1.1 validation, not
  Collection bookkeeping.

MutableCollection keeps only its public clearXsd11SchemaByNamespaceCache()
facade (GrammarTooling.java's external call site) and a single delegating
call from storeXmlDocument(); 9 imports that only the moved code needed are
dropped. Behavior-preserving: all 63 tests across the validation/XSD11 suite
pass unchanged.

Addresses dizzzz's review comments on #6530/#6551.
